Derive std::exception and throw it on false result
This fixes the error not throwing by explicitly checking that the result
is zero. According to the doc, `SUCCESS` is true only if the result is
non-negative. The inverse is true for `FAIL`, but in our case, zero is not
non-negative. We here explicitly check for 0.

Also, throwing with a string is not available nor is it recommended in the
event that the copy constructor for the std::string also throws its own
exception.